+++ common/web/input-processor/tests/cases/languageProcessor.js
@@ -8,9 +8,7 @@ import { Mock } from '@keymanapp/keyboard-processor';
  * Unit tests for the Dummy prediction model.
  */
 
-import kmlmc from '@keymanapp/lexical-model-compiler/dist/lexical-model-compiler/lexical-model-compiler.js';
-let LexicalModelCompiler = kmlmc.default; // Is compiled to CommonJS at present, which has interesting consequences.
-
+import { LexicalModelCompiler } from '@keymanapp/lexical-model-compiler';
 import { fileURLToPath } from 'url';
 import path from 'path';
 
@@ -53,7 +51,7 @@ describe('LanguageProcessor', function() {
     const __filename = fileURLToPath(import.meta.url);
     const __dirname = path.dirname(__filename);
 
-    const PATH = path.join(__dirname, '../../../../../developer/src/kmlmc/tests/fixtures', MODEL_ID);
+    const PATH = path.join(__dirname, '../../../../../developer/src/kmc-model/tests/fixtures', MODEL_ID);
 
     describe('using angle brackets for quotes', function() {
       let modelCode = compiler.generateLexicalModelCode(MODEL_ID, {