+++ common/web/lm-worker/build.sh
@@ -11,8 +11,8 @@ set -eu
 
 ## START STANDARD BUILD SCRIPT INCLUDE
 # adjust relative paths as necessary
-THIS_SCRIPT="$(greadlink -f "${BASH_SOURCE[0]}" 2>/dev/null || readlink -f "${BASH_SOURCE[0]}")"
-. "$(dirname "$THIS_SCRIPT")/../../../resources/build/build-utils.sh"
+THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
+. "${THIS_SCRIPT%/*}/../../../resources/build/build-utils.sh"
 ## END STANDARD BUILD SCRIPT INCLUDE
 
 . "$KEYMAN_ROOT/resources/shellHelperFunctions.sh"
@@ -54,7 +54,7 @@ fi
 
 if builder_start_action build; then
   # Build worker with tsc first
-  npm run build -- $builder_verbose || fail "Could not build worker."
+  npm run build -- $builder_verbose || builder_die "Could not build worker."
 
 
   echo "Bundling worker modules"