+++ developer/src/kmanalyze/kmanalyze.cpp
@@ -2,7 +2,6 @@
 //
 
 #include "pch.h"
-#include "../../../common/windows/cpp/include/crc32.h"
 #include "../../../common/windows/cpp/include/keymanversion.h"
 #include <iostream>
 #include <vector>
@@ -11,7 +10,6 @@
 
 BOOL LoadKeyboard(LPSTR fileName, LPKEYBOARD *lpKeyboard);
 
-BOOL VerifyChecksum(LPBYTE buf, LPDWORD CheckSum, DWORD sz);
 void Err(const char *p);
 int DoKeyboardAnalysis(LPKEYBOARD kbd, char *keyboardID, char *keyboardJSFilename, char *outputfilename);
 void MapVirtualKeys(void);
@@ -154,25 +152,21 @@ BOOL LoadKeyboard(LPSTR fileName, LPKEYBOARD *lpKeyboard)
   if (ckbp->dwFileVersion < VERSION_MIN ||
     ckbp->dwFileVersion > VERSION_MAX)
   {
-    /* Old or new version -- identify the desired program version */
-    if (VerifyChecksum(buf, &kbp->dwCheckSum, sz))
-    {
-      kbp->dpStoreArray = (LPSTORE)(buf + ckbp->dpStoreArray);
-      for (sp = kbp->dpStoreArray, i = 0; i < kbp->cxStoreArray; i++, sp++)
-        if (sp->dwSystemID == TSS_COMPILEDVERSION)
-        {
-          char buf2[256];
-          wsprintf(buf2, "Wrong File Version: file version is %ls", ((PBYTE)kbp) + (INT_PTR)sp->dpString);
-          delete buf;
-          Err(buf2);
-          return FALSE;
-        }
+    kbp->dpStoreArray = (LPSTORE)(buf + ckbp->dpStoreArray);
+    for (sp = kbp->dpStoreArray, i = 0; i < kbp->cxStoreArray; i++, sp++) {
+      if (sp->dwSystemID == TSS_COMPILEDVERSION)
+      {
+        char buf2[256];
+        wsprintf(buf2, "Wrong File Version: file version is %ls", ((PBYTE)kbp) + (INT_PTR)sp->dpString);
+        delete buf;
+        Err(buf2);
+        return FALSE;
+      }
     }
-    delete buf; Err("Unknown File Version: try using the latest version of KMDECOMP");
+    delete buf;
+    Err("Unknown File Version: try using the latest version of KMDECOMP");
     return FALSE;
   }
-
-  if (!VerifyChecksum(buf, &kbp->dwCheckSum, sz)) { delete buf; Err("Bad Checksum in file"); return FALSE; }
 
   kbp->dpStoreArray = (LPSTORE)(buf + ckbp->dpStoreArray);
   kbp->dpGroupArray = (LPGROUP)(buf + ckbp->dpGroupArray);
@@ -205,17 +199,6 @@ BOOL LoadKeyboard(LPSTR fileName, LPKEYBOARD *lpKeyboard)
   *lpKeyboard = kbp;
 
   return TRUE;
-}
-
-BOOL VerifyChecksum(LPBYTE buf, LPDWORD CheckSum, DWORD sz)
-{
-  DWORD tempcs;
-
-  tempcs = *CheckSum;
-  *CheckSum = 0;
-
-  BuildCRCTable();
-  return tempcs == CalculateBufferCRC(sz, buf);
 }