+++ resources/builder.inc.sh
@@ -30,8 +30,8 @@ function _builder_findRepoRoot() {
     # See https://stackoverflow.com/questions/59895/how-to-get-the-source-directory-of-a-bash-script-from-within-the-script-itself
     # None of the answers are 100% correct for cross-platform
     # On macOS, requires coreutils (`brew install coreutils`)
-    local SCRIPT=$(greadlink -f "${BASH_SOURCE[0]}" 2>/dev/null || readlink -f "${BASH_SOURCE[0]}")
-    REPO_ROOT=$(dirname $(dirname "$SCRIPT"))
+    local SCRIPT=$(readlink -f "${BASH_SOURCE[0]}")
+    REPO_ROOT="${SCRIPT%/*/*}"
     readonly REPO_ROOT
 }
 
@@ -44,22 +44,22 @@ function _builder_findRepoRoot() {
 # ```bash
 #   ## START STANDARD BUILD SCRIPT INCLUDE
 #   # adjust relative paths as necessary
-#   THIS_SCRIPT="$(greadlink -f "${BASH_SOURCE[0]}" 2>/dev/null || readlink -f "${BASH_SOURCE[0]}")"
-#   . "$(dirname "$THIS_SCRIPT")/resources/builder.inc.sh"
+#   THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
+#   . "${THIS_SCRIPT%/*}/resources/builder.inc.sh"
 #   ## END STANDARD BUILD SCRIPT INCLUDE
 # ```
 #
 function _builder_setBuildScriptIdentifiers() {
   if [ ! -z ${THIS_SCRIPT+x} ]; then
-    THIS_SCRIPT_PATH="$(dirname "$THIS_SCRIPT")"
+    THIS_SCRIPT_PATH="${THIS_SCRIPT%/*}"
     readonly THIS_SCRIPT_PATH
-    THIS_SCRIPT_NAME="$(basename "$THIS_SCRIPT")"
+    THIS_SCRIPT_NAME="${THIS_SCRIPT##*/}"
     readonly THIS_SCRIPT_NAME
     # Leaves only the part of the path based upon REPO_ROOT.
     THIS_SCRIPT_IDENTIFIER=${THIS_SCRIPT_PATH#"$REPO_ROOT/"}
     readonly THIS_SCRIPT_IDENTIFIER
   else
-    echo "Warning: THIS_SCRIPT not defined; builder.inc.sh has not been sourced with standard script include."
+    builder_die "THIS_SCRIPT not defined; builder.inc.sh has not been sourced with standard script include."
   fi
 }
 
@@ -83,10 +83,12 @@ builder_use_color() {
     COLOR_BLUE=$(tput setaf 4)
     COLOR_PURPLE=$(tput setaf 5)
     COLOR_TEAL=$(tput setaf 6)
-    COLOR_WHITE=$(tput setaf 7)
+    COLOR_WHITE=$(tput setaf 252)
+    COLOR_BRIGHT_WHITE=$(tput setaf 255)
     COLOR_GREY=$(tput setaf 8)
     COLOR_RESET=$(tput sgr0)
     # e.g. VSCode https://code.visualstudio.com/updates/v1_69#_setmark-sequence-support
+    BUILDER_BOLD=$(tput bold)
     HEADING_SETMARK='\x1b]1337;SetMark\x07'
 
     # Used by `builder_display_usage` when marking special terms (actions, targets, options)
@@ -101,8 +103,10 @@ builder_use_color() {
     COLOR_PURPLE=
     COLOR_TEAL=
     COLOR_WHITE=
+    COLOR_BRIGHT_WHITE=
     COLOR_GREY=
     COLOR_RESET=
+    BUILDER_BOLD=
     HEADING_SETMARK=
     BUILDER_TERM_START="<"
     BUILDER_TERM_END=">"
@@ -119,14 +123,23 @@ function builder_term() {
 
 function builder_die() {
   echo
-  echo "${COLOR_RED}$*${COLOR_RESET}"
+  if [[ $# -eq 0 ]]; then
+    builder_echo error "Unspecified error, aborting script"
+  else
+    builder_echo error "$*"
+  fi
   echo
   exit 1
 }
 
 function builder_warn() {
-  echo "${COLOR_YELLOW}$*${COLOR_RESET}"
+  builder_echo warning "$*"
 }
+
+function builder_heading() {
+  builder_echo heading "$*"
+}
+
 
 ####################################################################################
 #
@@ -134,17 +147,49 @@ function builder_warn() {
 #
 ####################################################################################
 
+
+builder_echo() {
+  local color=white message= mark=
+  if [[ $# -gt 1 ]]; then
+    color="$1"
+    message="$2"
+  else
+    message="$1"
+  fi
+
+  if [[ ! -z ${COLOR_RED+x} ]]; then
+    case $color in
+      white) color="$COLOR_WHITE" ;;
+      grey) color="$COLOR_GREY" ;;
+      green|success) color="$COLOR_GREEN" ;;
+      blue|heading) color="$COLOR_BLUE" ;;
+      yellow|warning) color="$COLOR_YELLOW" ;;
+      red|error) color="$COLOR_RED" ;;
+      purple) color="$COLOR_PURPLE" ;;
+      brightwhite) color="$COLOR_BRIGHTWHITE" ;;
+      teal|debug) color="$COLOR_TEAL" ;;
+      setmark) mark="$HEADING_SETMARK" color="$COLOR_PURPLE" ;;
+    esac
+
+    if builder_is_dep_build; then
+      echo -e "$mark$COLOR_GREY[$THIS_SCRIPT_IDENTIFIER]$COLOR_RESET $color$message$COLOR_RESET"
+    else
+      echo -e "$mark$BUILDER_BOLD$COLOR_BRIGHT_WHITE[$THIS_SCRIPT_IDENTIFIER]$COLOR_RESET $color$message$COLOR_RESET"
+    fi
+  else
+    # Cope with the case of pre-init message and just emit plain text
+    echo -e "$message"
+  fi
+}
+
+builder_echo_debug() {
+  builder_echo debug "[DEBUG] $*"
+}
+
 #
 # builder_ names are reserved.
 # _builder_ names are internal use and subject to change
 #
-if [ -z ${_builder_debug+x} ]; then
-  _builder_debug=false
-fi
-
-if $_builder_debug; then
-  echo "[DEBUG] Command line: $0 $@"
-fi
 
 #
 # builder_extra_params: string containing all parameters after '--'
@@ -199,8 +244,7 @@ _builder_item_is_target() {
 
 function _builder_warn_if_incomplete() {
   if [ -n "${_builder_current_action}" ]; then
-    local scope="[$THIS_SCRIPT_IDENTIFIER] "
-    echo "${COLOR_YELLOW}## ${scope}Warning - $_builder_current_action never reported success or failure${COLOR_RESET}"
+    builder_echo warning "$_builder_current_action never reported success or failure"
     # exit 1  # If we wanted this scenario to result in a forced build-script fail.
   fi
 
@@ -250,8 +294,8 @@ _builder_failure_trap() {
 #
 _builder_cleanup_deps() {
   if ! builder_is_dep_build && [[ ! -z ${_builder_deps_built+x} ]]; then
-    if $_builder_debug; then
-      echo "[DEBUG] Dependencies that were built:"
+    if $_builder_debug_internal; then
+      builder_echo_debug "Dependencies that were built:"
       cat "$_builder_deps_built"
     fi
     rm -f "$_builder_deps_built"
@@ -259,29 +303,42 @@ _builder_cleanup_deps() {
   fi
 }
 
+#------------------------------------------------------------------------------------------
+# Child scripts
+#------------------------------------------------------------------------------------------
+
 _builder_execute_child() {
   local action=$1
   local target=$2
 
-  local scope="[$THIS_SCRIPT_IDENTIFIER] "
   local script="$THIS_SCRIPT_PATH/${_builder_target_paths[$target]}/build.sh"
 
-  if $_builder_debug; then
-    echo "${COLOR_BLUE}## $scope$action$target starting...${COLOR_RESET}"
+  if $_builder_debug_internal; then
+    builder_echo heading "## $action$target starting..."
   fi
 
+  # Build array of specified inheritable options
+  local child_options=()
+  local opt
+  for opt in "${_builder_options_inheritable[@]}"; do
+    if builder_has_option $opt; then
+      child_options+=($opt)
+    fi
+  done
+
   "$script" $action \
+    ${child_options[@]} \
     $builder_verbose \
     $builder_debug \
   && (
-    if $_builder_debug; then
-      echo "${COLOR_GREEN}## $scope$action$target completed successfully${COLOR_RESET}"
+    if $_builder_debug_internal; then
+      builder_echo success "## $action$target completed successfully"
     fi
   ) || (
     result=$?
-    echo "${COLOR_RED}## $scope$action$target failed with exit code $result${COLOR_RESET}"
+    builder_echo error "## $action$target failed with exit code $result"
     exit $result
-  )
+  ) || exit $? # Required due to above subshell masking exit
 }
 
 _builder_run_child_action() {
@@ -342,6 +399,10 @@ builder_run_child_actions() {
     shift
   done
 }
+
+#------------------------------------------------------------------------------------------
+# Various API endpoints
+#------------------------------------------------------------------------------------------
 
 #
 # Builds the standardized `action:target` string for the specified action-target
@@ -416,21 +477,18 @@ builder_has_action() {
 # ```
 #
 builder_start_action() {
-  local scope="[$THIS_SCRIPT_IDENTIFIER] "
-
   if builder_has_action $1; then
     # In a dependency quick build (the default), determine whether we actually
     # need to run this step. Uses data passed to builder_describe_outputs to
     # verify whether a target output is present.
     if builder_is_dep_build &&
         ! builder_is_full_dep_build &&
-        [[ ! -z ${_builder_dep_path[$_builder_matched_action]+x} ]] &&
-        [[ -e "$KEYMAN_ROOT/${_builder_dep_path[$_builder_matched_action]}" ]]; then
-      echo "$scope skipping $_builder_matched_action_name, up-to-date"
+        _builder_dep_output_exists $_builder_matched_action; then
+      builder_echo "skipping $_builder_matched_action_name, up-to-date"
       return 1
     fi
 
-    echo "${COLOR_BLUE}## $scope$_builder_matched_action_name starting...${COLOR_RESET}"
+    builder_echo blue "## $_builder_matched_action_name starting..."
     if [ -n "${_builder_current_action}" ]; then
       _builder_warn_if_incomplete
     fi
@@ -549,16 +607,21 @@ _builder_expand_action_targets() {
 #
 # There are four types of parameters that may be specified:
 #
-# * **Option:** `"--option[,-o][=var]   [One line description]"`
+# * **Option:** `"--option[,-o][+][=var]   [One line description]"`
 #
 #   All options must have a longhand form with two prefix hyphens,
 #   e.g. `--option`. The `,-o` shorthand form is optional. When testing if
 #   the option is set with `builder_has_option`, always use the longhand
 #   form.
 #
-#   if `=var` is specified, then the next parameter will be a variable stored in
+#   If `=var` is specified, then the next parameter will be a variable stored in
 #   `$var` for that option. e.g. `--option=opt` means `$opt` will have the value
 #   `"foo"` when the script is called for `--option foo`.
+#
+#   If `+` is specified, then the option will be passed to child scripts. All
+#   child scripts _must_ accept this option, or they will fail. It is acceptable
+#   for the child script to declare the option but ignore it. However, the option
+#   will _not_ be passed to dependencies.
 #
 # * **Action**: `"action   [One line description]"`
 #
@@ -577,7 +640,7 @@ _builder_expand_action_targets() {
 #   `=path` to the target definition, for example `:app=src/app`. Where
 #   possible, avoid differences in names of child projects and folders.
 #
-# * **Dependency:** "@/path/to/dependency [action][:target] ..."
+# * **Dependency:** `"@/path/to/dependency[:target] [action][:target] ..."``
 #
 #   A dependency always starts with `@`. The path to the dependency will be
 #   relative to the build script folder, or to the root of the repository, if
@@ -587,10 +650,14 @@ _builder_expand_action_targets() {
 #   Relative paths will be expanded to full paths, again, relative to the root
 #   of the repository.
 #
-#   Dependencies may be limited to specific `action:target`. If not specified,
-#   dependencies will be built for all actions on all targets. Either `action`
-#   or `:target` may be omitted, and multiple actions and targets may be
-#   specified, space separated.
+#   A dependency definition can include a target for that dependency, for
+#   example, `"@/core:arch"`. This would build only the ':arch' target for the
+#   core module.
+#
+#   Dependencies may be limited to specific `action:target` pairs on the current
+#   script. If not specified, dependencies will be built for all actions on all
+#   targets. Either `action` or `:target` may be omitted, and multiple actions
+#   and targets may be specified, space separated.
 #
 builder_describe() {
   _builder_record_function_call builder_describe
@@ -600,6 +667,7 @@ builder_describe() {
   _builder_targets=()
   _builder_options=()
   _builder_deps=()                    # array of all dependencies for this script
+  _builder_options_inheritable=()     # array of all options that should be passed to child scripts
   _builder_default_action=build
   declare -A -g _builder_params
   declare -A -g _builder_options_short
@@ -608,6 +676,7 @@ builder_describe() {
   declare -A -g _builder_dep_related_actions  # array of action:targets associated with a given dependency
   declare -A -g _builder_internal_dep         # array of internal action:targets dependency relationships
   declare -A -g _builder_target_paths         # array of target child project paths
+  declare -A -g _builder_dep_targets          # array of :targets given for a specific dependency (comma separated if more than one)
   shift
   # describe each target, action, and option possibility
   while [[ $# -gt 0 ]]; do
@@ -642,10 +711,16 @@ builder_describe() {
     elif [[ $value =~ ^@ ]]; then
       # Parameter is a dependency
       local dependency="${value:1}"
+      local dependency_target= # all targets
+      if [[ $dependency =~ : ]]; then
+        dependency_target=":$(echo "$dependency" | cut -d: -f 2 -)"
+        dependency="$(echo "$dependency" | cut -d: -f 1 -)"
+      fi
+
       dependency="`_builder_expand_relative_path "$dependency"`"
       _builder_deps+=($dependency)
       _builder_dep_related_actions[$dependency]="`_builder_expand_action_targets "$description"`"
-
+      _builder_dep_targets[$dependency]="$dependency_target"
       # We don't want to add deps to params, so shift+continue
       shift
       continue
@@ -658,10 +733,21 @@ builder_describe() {
         value="$(echo "$value" | cut -d= -f 1 -)"
       fi
 
+      local is_inheritable=false
+
+      if [[ $value =~ \+$ ]]; then
+        # final + indicates that option is inheritable
+        is_inheritable=true
+        value="${value:0:-1}"
+      fi
+
       if [[ $value =~ , ]]; then
         local option_long="$(echo "$value" | cut -d, -f 1 -)"
         local option_short="$(echo "$value" | cut -d, -f 2 -)"
         _builder_options+=($option_long)
+        if $is_inheritable; then
+          _builder_options_inheritable+=($option_long)
+        fi
         _builder_options_short[$option_short]="$option_long"
         if [[ ! -z "$option_var" ]]; then
           _builder_options_var[$option_long]="$option_var"
@@ -669,6 +755,9 @@ builder_describe() {
         value="$option_long, $option_short"
       else
         _builder_options+=($value)
+        if $is_inheritable; then
+          _builder_options_inheritable+=($value)
+        fi
         if [[ ! -z "$option_var" ]]; then
           _builder_options_var[$value]="$option_var"
         fi
@@ -677,6 +766,7 @@ builder_describe() {
       if [[ ! -z $option_var ]]; then
         value="$value $option_var"
       fi
+
     else
       # Parameter is an action
       if [[ $value =~ \+$ ]]; then
@@ -726,8 +816,8 @@ builder_describe() {
 #
 # ```bash
 #   builder_describe_outputs \
-#     configure /node_modules \
-#     build     build/index.js
+#     "configure" "/node_modules" \
+#     "build"     "build/index.js"
 # ```
 #
 function builder_describe_outputs() {
@@ -784,11 +874,10 @@ _builder_parameter_error() {
   local program="$1"
   local type="$2"
   local param="$3"
-  echo "$COLOR_RED$program: invalid $type: $param$COLOR_RESET"
+  builder_echo red "$program: invalid $type: $param"
   echo
   builder_display_usage
   exit 64
-
 }
 
 #
@@ -820,7 +909,7 @@ _builder_check_color() {
 # its full internal dependency tree
 #
 _builder_add_chosen_action_target_dependencies() {
-  local action_target e i=0 new_actions=()
+  local action_target i=0 new_actions=()
 
   # Iterate through every action specified on command line; we use this loop
   # style so that any new actions added here will also be iteratively checked
@@ -829,30 +918,32 @@ _builder_add_chosen_action_target_dependencies() {
 
     # If we have an internal dependency for the chosen action:target pair
     if [[ ! -z ${_builder_internal_dep[$action_target]+x} ]]; then
-      local dep_output=${_builder_internal_dep[$action_target]}
-      # If there is a defined output for this dependency
-      if [[ ! -z ${_builder_dep_path[$dep_output]+x} ]]; then
-        # If the output for the dependency is missing, or we have --force-deps
-        if [[ ! -e "$KEYMAN_ROOT/${_builder_dep_path[$dep_output]}" ]] || builder_is_full_dep_build; then
-          # Add the dependency to the chosen action:target list
-          if ! _builder_item_in_array "$dep_output" "${_builder_chosen_action_targets[@]}"; then
-            _builder_chosen_action_targets+=($dep_output)
-            new_actions+=($dep_output)
+      local dep_outputs=(${_builder_internal_dep[$action_target]}) dep_output
+      for dep_output in "${dep_outputs[@]}"; do
+        # If there is a defined output for this dependency
+        if [[ ! -z ${_builder_dep_path[$dep_output]+x} ]]; then
+          # If the output for the dependency is missing, or we have --force-deps
+          if [[ ! -e "$KEYMAN_ROOT/${_builder_dep_path[$dep_output]}" ]] || builder_is_full_dep_build; then
+            # Add the dependency to the chosen action:target list
+            if ! _builder_item_in_array "$dep_output" "${_builder_chosen_action_targets[@]}"; then
+              _builder_chosen_action_targets+=($dep_output)
+              new_actions+=($dep_output)
+            fi
           fi
         fi
-      fi
+      done
     fi
     i=$((i + 1))
   done
 
   if [[ ${#new_actions[@]} -gt 0 ]]; then
     if builder_is_full_dep_build; then
-      echo "Automatically running all dependency actions due to --force-deps:"
+      builder_echo "Automatically running all dependency actions due to --force-deps:"
     else
-      echo "Automatically running following required actions with missing outputs:"
+      builder_echo "Automatically running following required actions with missing outputs:"
     fi
     for e in "${new_actions[@]}"; do
-      echo "* $e"
+      builder_echo "* $e"
     done
   fi
 }
@@ -882,8 +973,37 @@ _builder_define_default_internal_dep() {
   local target=$1 dep=$2 action=$3
   if _builder_item_in_array $dep "${_builder_actions[@]}" &&
         _builder_item_in_array $action "${_builder_actions[@]}"; then
-    _builder_internal_dep[$action$target]=$dep$target
+    [[ -z ${_builder_internal_dep[$action$target]+x} ]] &&
+      _builder_internal_dep[$action$target]=$dep$target ||
+      _builder_internal_dep[$action$target]="${_builder_internal_dep[$action$target]} $dep$target"
   fi
+}
+
+#
+# Define a local dependency between one action:target and
+# another.
+#
+# Usage:
+#   builder_describe_internal_dependency action:target depaction:deptarget ...
+# Parameters:
+#   1: action:target         The action and target that has a dependency
+#   2: depaction:deptarget   The dependency action and target
+# Example:
+#   builder_describe_internal_dependency \
+#     mac:build mac-x86_64:build \
+#     mac:build mac-arm64:build
+#
+# Note: actions and targets must be fully specified, and this _must_
+# be called before either builder_describe_outputs or builder_parse in
+# order for dependencies to be resolved.
+builder_describe_internal_dependency() {
+  while [[ $# -gt 0 ]]; do
+    local action_target=$1 dep_action_target=$2
+    [[ -z ${_builder_internal_dep[$action_target]+x} ]] &&
+      _builder_internal_dep[$action_target]=$dep_action_target ||
+      _builder_internal_dep[$action_target]="${_builder_internal_dep[$action_target]} $dep_action_target"
+    shift 2
+  done
 }
 
 # Initializes a build.sh script, parses command line. Will abort the script if
@@ -897,6 +1017,8 @@ _builder_define_default_internal_dep() {
 builder_parse() {
 
   _builder_record_function_call builder_parse
+
+  local _builder_params="$@"
 
   _builder_build_deps=--deps
   builder_verbose=
@@ -1027,28 +1149,29 @@ builder_parse() {
     _builder_add_chosen_action_target_dependencies
   fi
 
-  if $_builder_debug; then
-    echo "[DEBUG] Selected actions and targets:"
+  if $_builder_debug_internal; then
+    builder_echo_debug "Selected actions and targets:"
     for e in "${_builder_chosen_action_targets[@]}"; do
-      echo "* $e"
+      builder_echo_debug "* $e"
     done
-    echo
-    echo "[DEBUG] Selected options:"
+    builder_echo_debug
+    builder_echo_debug "Selected options:"
     for e in "${_builder_chosen_options[@]}"; do
-      echo "* $e"
+      builder_echo_debug "* $e"
     done
   fi
 
   if builder_is_dep_build; then
-    echo "[$THIS_SCRIPT_IDENTIFIER] dependency build, started by $builder_dep_parent"
+    builder_echo setmark "dependency build, started by $builder_dep_parent"
+    builder_echo grey "build.sh parameters: <${_builder_params[@]}>"
     if [[ -z ${_builder_deps_built+x} ]]; then
-      echo "FATAL ERROR: Expected --builder-deps-built parameter"
-      exit 1
+      builder_die "FATAL ERROR: Expected --builder-deps-built parameter"
     fi
   else
     # This is a top-level invocation, not a dependency build, so we want to
     # track which dependencies have been built, so they don't get built multiple
     # times.
+    builder_echo setmark "build.sh parameters: <${_builder_params[@]}>"
     _builder_deps_built=`mktemp`
   fi
 
@@ -1101,7 +1224,7 @@ builder_display_usage() {
   echo "Actions: "
 
   for e in "${_builder_actions[@]}"; do
-    if [[ -v _builder_params[$e] ]]; then
+    if _builder_item_in_glob_array "$e" "${_builder_params[@]}"; then
       description="${_builder_params[$e]}"
     else
       description=$(_builder_get_default_description "$e")
@@ -1113,7 +1236,7 @@ builder_display_usage() {
   echo "Targets: "
 
   for e in "${_builder_targets[@]}"; do
-    if [[ -v _builder_params[$e] ]]; then
+    if _builder_item_in_glob_array "$e" "${_builder_params[@]}"; then
       description="${_builder_params[$e]}"
     else
       description=$(_builder_get_default_description "$e")
@@ -1180,21 +1303,48 @@ builder_finish_action() {
     action_name="$action$target"
   fi
 
-  local scope="[$THIS_SCRIPT_IDENTIFIER] "
+  local matched_action="$action$target"
 
-  if [[ "$action$target" == "${_builder_current_action}" ]]; then
+  if [[ "$matched_action" == "${_builder_current_action}" ]]; then
     if [[ $result == success ]]; then
-      echo "${COLOR_GREEN}## $scope$action_name completed successfully${COLOR_RESET}"
+      # Sanity check:  if there is a described output for this action, does the corresponding
+      # file or directory exist now?
+      if _builder_dep_output_defined $matched_action && ! _builder_dep_output_exists "$matched_action"; then
+        builder_echo warning "Expected output: '${_builder_dep_path[$matched_action]}'."
+        builder_echo warning "## $action_name completed successfully, but output does not exist"
+      else
+        builder_echo success "## $action_name completed successfully"
+      fi
     elif [[ $result == failure ]]; then
-      echo "${COLOR_RED}## $scope$action_name failed${COLOR_RESET}"
+      builder_echo error "## $action_name failed"
     else
-      echo "${COLOR_RED}## $scope$action_name failed with message: $result${COLOR_RESET}"
+      builder_echo error "## $action_name failed with message: $result"
     fi
 
     # Remove $action$target from the array; it is no longer a current action
     _builder_current_action=
   else
-    echo "${COLOR_YELLOW}## Warning: reporting result of $action_name but the action was never started!${COLOR_RESET}"
+    builder_echo warning "reporting result of $action_name but the action was never started!"
+  fi
+}
+
+#------------------------------------------------------------------------------------------
+# Dependencies
+#------------------------------------------------------------------------------------------
+
+_builder_dep_output_defined() {
+  if [[ ! -z ${_builder_dep_path[$1]+x} ]]; then
+    return 0
+  else
+    return 1
+  fi
+}
+
+_builder_dep_output_exists() {
+  if _builder_dep_output_defined $1 && [[ -e "$KEYMAN_ROOT/${_builder_dep_path[$1]}" ]]; then
+    return 0
+  else
+    return 1
   fi
 }
 
@@ -1204,15 +1354,38 @@ builder_finish_action() {
 _builder_should_build_dep() {
   local action_target="$1"
   local dep="$2"
+  echo $dep
   local related_actions=(${_builder_dep_related_actions[$dep]})
-  # echo "bdra: ${_builder_dep_related_actions[@]}"
-  # echo "target: $action_target"
-  # echo "dep: $2"
-  # echo "ra: ${related_actions[@]}"
+
+  if [[ $action_target =~ ^clean ]]; then
+    # don't attempt to build dependencies for a 'clean' action
+    return 1
+  fi
+
   if ! _builder_item_in_glob_array "$action_target" "${related_actions[@]}"; then
     return 1
   fi
   return 0
+}
+
+#
+# Removes a dependency from the list of available dependencies
+#
+# Parameters:
+#   $1    path to dependency
+#
+builder_remove_dep() {
+  local dependency="$1" i
+  dependency="`_builder_expand_relative_path "$dependency"`"
+
+  for i in "${!_builder_deps[@]}"; do
+    if [[ ${_builder_deps[i]} = $dependency ]]; then
+      unset '_builder_deps[i]'
+    fi
+  done
+
+  # rebuild the array to remove the empty item
+  _builder_deps=( "${_builder_deps[@]}" )
 }
 
 #
@@ -1231,7 +1404,7 @@ _builder_do_build_deps() {
     # Don't attempt to build dependencies that don't match the current
     # action:target (wildcards supported for matches here)
     if ! _builder_should_build_dep "$action_target" "$dep"; then
-      echo "[$THIS_SCRIPT_IDENTIFIER] Skipping dependency build $dep for $_builder_matched_action_name"
+      builder_echo "Skipping dependency $dep for $_builder_matched_action_name"
       continue
     fi
 
@@ -1240,14 +1413,28 @@ _builder_do_build_deps() {
       continue
     fi
 
-    # TODO: add --debug as a standard builder parameter
+    dep_target=
+    if [[ ! -z ${_builder_dep_targets[$dep]+x} ]]; then
+      # TODO: in the future split _builder_dep_targets into comma-separated
+      #       array for multiple targets for a dep?
+      dep_target=${_builder_dep_targets[$dep]}
+    fi
+
     builder_set_module_has_been_built "$dep"
-    "$KEYMAN_ROOT/$dep/build.sh" configure build \
+    "$KEYMAN_ROOT/$dep/build.sh" "configure$dep_target" "build$dep_target" \
       $builder_verbose \
       $builder_debug \
       $_builder_build_deps \
       --builder-deps-built "$_builder_deps_built" \
-      --builder-dep-parent "$THIS_SCRIPT_IDENTIFIER"
+      --builder-dep-parent "$THIS_SCRIPT_IDENTIFIER" && (
+      if $_builder_debug_internal; then
+        builder_echo success "## Dependency $dep for $_builder_matched_action_name successfully"
+      fi
+    ) || (
+      result=$?
+      builder_echo error "## Dependency failed with exit code $result"
+      exit $result
+    ) || exit $? # Required due to above subshell masking exit
   done
 }
 
@@ -1367,6 +1554,19 @@ builder_set_module_has_been_built() {
 }
 
 #
+# Reports on all described dependencies, then exits
+# used by builder-controls.sh
+#
+_builder_report_dependencies() {
+  echo "${_builder_deps[@]}"
+  exit 0
+}
+
+#------------------------------------------------------------------------------------------
+# Utility functions
+#------------------------------------------------------------------------------------------
+
+#
 # returns `0` if we should be verbose in output
 #
 builder_verbose() {
@@ -1379,20 +1579,11 @@ builder_verbose() {
 #
 # returns `0` if we are doing a debug build
 #
-builder_debug() {
+builder_is_debug_build() {
   if [[ $builder_debug == --debug ]]; then
     return 0
   fi
   return 1
-}
-
-#
-# Reports on all described dependencies, then exits
-# used by builder-controls.sh
-#
-_builder_report_dependencies() {
-  echo "${_builder_deps[@]}"
-  exit 0
 }
 
 #
@@ -1425,3 +1616,15 @@ _builder_has_function_been_called() {
 #
 _builder_init
 _builder_check_color "$@"
+
+# _builder_debug_internal flag can be used to emit verbose logs for builder itself,
+# e.g.:
+#   _builder_debug_internal=true ./build.sh
+#
+if [ -z ${_builder_debug_internal+x} ]; then
+  _builder_debug_internal=false
+fi
+
+if $_builder_debug_internal; then
+  builder_echo_debug "Command line: $0 $@"
+fi